remove fs from mocha tests
