+++ packages/datadog-plugin-mocha/test/index.spec.js
@@ -92,7 +92,7 @@ describe('Plugin', () => {
         .get('/')
         .reply(200, 'OK')
 
-      return agent.load(['mocha', 'fs', 'http']).then(() => {
+      return agent.load(['mocha', 'http']).then(() => {
         Mocha = require(`../../../versions/mocha@${version}`).get()
       })
     })
@@ -256,27 +256,6 @@ describe('Plugin', () => {
 
       it('works with integrations', (done) => {
         const testFilePath = path.join(__dirname, 'mocha-test-integration.js')
-
-        agent.use(trace => {
-          const testSpan = trace[0].find(span => span.type === 'test')
-          const fsOperationSpan = trace[0].find(span => span.name === 'fs.operation')
-          expect(testSpan.parent_id.toString()).to.equal('0')
-          expect(testSpan.meta[ORIGIN_KEY]).to.equal(CI_APP_ORIGIN)
-          expect(testSpan.meta[TEST_STATUS]).to.equal('pass')
-          expect(testSpan.meta[TEST_NAME]).to.equal('mocha-test-integration can do integration tests')
-          expect(fsOperationSpan.parent_id.toString()).to.equal(testSpan.span_id.toString())
-          expect(fsOperationSpan.meta[ORIGIN_KEY]).to.equal(CI_APP_ORIGIN)
-        }).then(done, done)
-
-        const mocha = new Mocha({
-          reporter: function () {} // silent on internal tests
-        })
-        mocha.addFile(testFilePath)
-        mocha.run()
-      })
-
-      it('works with http integration', (done) => {
-        const testFilePath = path.join(__dirname, 'mocha-test-integration-http.js')
         const testSuite = testFilePath.replace(`${process.cwd()}/`, '')
 
         agent.use(trace => {
